REF: Consolidate related API code to flatten and denest _rpc and dbapi submodules. flake8 cleaning
